User Documentation

ExaPowerIO exports three functions:

ExaPowerIO.parse_pglib — Function

function parse_pglib(
    ::Type{T<:Real},
    dataset_query :: String;
    out_type=PowerData{T}
) :: Union{PowerData{T}, NamedTuple} where T

Searches the PGLib database for dataset_query, and errors if there is not exactly 1 result. Otherwise, the matching file is downloaded and parsed to return a value of type out_type.

Each value will be parsed as a T.

Currently, out_type can only be ExaPowerIO.PowerData{T}, or NamedTuple.

parse_file(dataset_query; out_file) will return parse_file(Float64, dataset_query; out_file)

ExaPowerIO.parse_file — Function

parse_file(
    fname :: String;
    out_type=PowerData{T}
)

Parses the Matpower file specified by fname, and returns a value of type out_type.

Each value will be parsed as a T.

Currently, out_type can only be ExaPowerIO.PowerData{T}, or NamedTuple.

parse_file(dataset_query; out_file) will return parse_file{Float64}(dataset_query; out_file)

ExaPowerIO.PowerData — Type

struct Data{T <: Real}
    version :: String
    baseMVA :: T
    bus :: Vector{BusData{T}}
    gen :: Vector{GenData{T}}
    branch :: Vector{BranchData{T}}
    storage :: Vector{StorageData{T}}
end

version, baseMVA, bus, gen, branch, and storage all corespond to members of the mpc object created by a matpower file. Their fields correspond exactly with the columns of the relevant mpc member.

ExaPowerIO.BusData — Type

struct BusData{T <: Real}
    bus_i :: Int
    type :: Int
    pd :: T
    qd :: T
    gs :: T
    bs :: T
    area :: Int
    vm :: T
    va :: T
    baseKV :: T
    zone :: Int
    vmax :: T
    vmin :: T
end

ExaPowerIO.GenData — Type

struct GenData{T <: Real}
    bus :: Int
    pg :: T
    qg :: T
    qmax :: T
    qmin :: T
    vg :: T
    mbase :: T
    status :: Int
    pmax :: T
    pmin :: T
    i :: Int
    model_poly :: Bool
    startup :: T
    shutdown :: T
    n :: Int
    c :: NTuple{3, T}
end

bus is an index into the PowerData.bus Vector, not bus_i values

ExaPowerIO.BranchData — Type

struct BranchData{T <: Real}
    fbus :: Int
    tbus :: Int
    br_r :: T
    br_x :: T
    b_fr :: T,
    b_to :: T,
    g_fr :: T,
    g_to :: T,
    ratea ::T
    rateb :: T
    ratec :: T
    tap :: T
    shift :: T
    status :: Int
    angmin :: T
    angmax :: T
    c1 :: T
    c2 :: T
    c3 :: T
    c4 :: T
    c5 :: T
    c6 :: T
    c7 :: T
    c8 :: T
end

fbus and tbus are indices into the PowerData.bus Vector, not bus_i values

ExaPowerIO.StorageData — Type

struct StorageData{T <: Real}
    storage_bus :: T
    ps :: Int
    qs :: T
    energy :: T
    energy_rating :: T
    charge_rating :: T
    discharge_rating :: T
    charge_efficiency :: T
    discharge_efficiency :: T
    thermal_rating :: T
    qmin :: T
    qmax :: T
    r :: T
    x :: T
    p_loss :: T
    q_loss :: T
    status :: Int
end